requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
rembg==2.0.56
pillow==10.3.0
python-multipart==0.0.9
